yolo-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
ultralytics==8.3.55
easyocr==1.7.2
Pillow==11.1.0
numpy==1.26.4
python-multipart==0.0.20
